+++ src/node/index.js
@@ -1,5 +1,6 @@
 import { app, ipcMain as ipc, dialog, BrowserWindow } from "electron";
 import debug from "electron-debug";
+import isDev from "electron-is-dev";
 import { join } from "path";
 import { cleanFile, cleanArray, readFiles } from "./utils";
 
@@ -15,7 +16,11 @@ app.on("window-all-closed", () => {
 app.on("ready", () => {
   mainWindow = new BrowserWindow({width: 500, height: 600});
 
-  mainWindow.loadURL(`file://${join(process.cwd(), "index.html")}`);
+  if(isDev) {
+    mainWindow.loadURL(`file://${join(process.cwd(), "index.html")}`);
+  } else {
+    mainWindow.loadURL(`file://${__dirname}/index.html`);
+  }
 
   mainWindow.on("closed", () => {
     mainWindow = null;